requirements.txt
asgiref==3.2.3
betterpath==0.2.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==3.0
django-allauth==0.40.0
django-cors-headers==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
l==0.11.0
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vcversioner==2.16.0.0
zope.interface==4.7.1
